# api/openapi.yaml
openapi: 3.0.3
info:
  title: Helling API
  version: 0.2.0
  description: |
    Helling is a self-hosted, single-node container and VM management platform.
    This is the v0.2 Platform Core bridge: authentication, instance CRUD,
    operation polling, schedules, webhooks, host firewall, events, audit, and
    user Incus scope assignment.

    All endpoints under `/v1/` are stable within v0.2. Breaking changes ship under `/v2/`.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Bizarre Industries
    url: https://github.com/Bizarre-Industries/helling
servers:
  - url: https://localhost/v1
    description: Local installation behind the Caddy edge service
  - url: http+unix:///run/helling/api.sock/v1
    description: Direct socket access (CLI, dev)
...
security:
  - cookieAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List local users
      description: Returns all local Helling users. Admin role required.
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListEnvelope"
              example:
                data:
                  - id: 1
                    username: admin
                    is_admin: true
                    incus_project: default
                    created_at: "2026-05-03T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags:
        - users
      summary: Create a local user
      description: Creates a Helling user and provisions its delegated Incus scope. Admin role required.
      requestBody:
        required: true
        description: User creation request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
            example:
              username: alice
              password: correct horse battery staple
              is_admin: false
              incus_project: alice
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                data:
                  id: 2
                  username: alice
                  is_admin: false
                  incus_project: alice
                  created_at: "2026-05-05T09:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      operationId: getUser
      tags:
        - users
      summary: Get a local user
      description: Returns one Helling user. Admin role required.
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                data:
                  id: 2
                  username: alice
                  is_admin: false
                  incus_project: alice
                  created_at: "2026-05-05T09:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateUser
      tags:
        - users
      summary: Update a local user
      description: Partially updates mutable user fields. Admin role required.
      requestBody:
        required: true
        description: Partial user update request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
            example:
              is_admin: false
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                data:
                  id: 2
                  username: alice
                  is_admin: false
                  incus_project: alice
                  created_at: "2026-05-05T09:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUser
      tags:
        - users
      summary: Delete a local user
      description: Deletes a user and revokes its delegated Incus trust entry. Admin role required.
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: helling_session
      description: Session cookie issued by `POST /auth/login`. Sent automatically by the browser; CLI stores it under `~/.config/helling/session`.
  parameters:
...
    UserID:
      in: path
      name: id
      required: true
      description: Local user id path parameter.
      schema:
        type: integer
        description: Stable local user id.
        format: int64
        minimum: 1
        example: 2
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated principal lacks the required role or token scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State conflict (e.g. instance already exists, already running)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    User:
      type: object
      description: Authenticated user identity.
      required: [id, username, is_admin, created_at]
      properties:
        id:
          type: integer
          description: Stable user identifier.
          format: int64
          example: 1
        username:
          type: string
          description: Account username.
          example: alice
        is_admin:
          type: boolean
          description: True for the bootstrap admin user.
          example: false
        incus_project:
          type: string
          description: Incus project assigned to this user for delegated proxy calls.
          example: alice
        created_at:
          type: string
          description: Account creation time (RFC 3339).
          format: date-time
          example: "2026-05-03T12:00:00Z"
...
    UserEnvelope:
      type: object
      description: Envelope containing one user.
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/User"
    UserListEnvelope:
      type: object
      description: Envelope containing users.
      required: [data]
      properties:
        data:
          type: array
          description: User payload list.
          example:
            - id: 1
              username: admin
              is_admin: true
              incus_project: default
              created_at: "2026-05-03T12:00:00Z"
          items:
            $ref: "#/components/schemas/User"
    UserCreateRequest:
      type: object
      description: Request body for creating a local user.
      required: [username, password]
      properties:
        username:
          type: string
          description: Account username.
          minLength: 1
          maxLength: 32
          example: alice
        password:
          type: string
          description: Initial account password.
          minLength: 8
          maxLength: 1024
          format: password
          writeOnly: true
          example: correct horse battery staple
        is_admin:
          type: boolean
          description: True when the created user should receive admin role.
          default: false
          example: false
        incus_project:
          type: string
          description: Incus project assigned to the user's restricted certificate.
          minLength: 1
          maxLength: 63
          example: alice
    UserUpdateRequest:
      type: object
      description: Partial request body for updating a local user.
      properties:
        password:
          type: string
          description: Replacement account password.
          minLength: 8
          maxLength: 1024
          format: password
          writeOnly: true
          example: correct horse battery staple
        is_admin:
          type: boolean
          description: Replacement admin-role flag.
          example: false
...
    Error:
      type: object
      description: Standard error envelope returned by 4xx and 5xx responses.
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: not_found
        message:
          type: string
          description: Human-readable message.
          example: instance not found
        details:
          type: object
          description: Optional structured context.
          additionalProperties: true
          example:
            field: name